Move phaser defaults out of startup.js
