+++ src/core/loader/container.js
@@ -1,0 +1,10 @@
+import { gmi } from "../gmi/gmi.js";
+
+export const getContainerDiv = () =>{
+    const containerDiv = document.getElementById(gmi.gameContainerId);
+    if (!containerDiv) {
+        throw Error(`Container element "#${gmi.gameContainerId}" not found`);
+    } else {
+        return containerDiv;
+    }
+}
